+++ src/main/java/com/donate/donate_app/entity/Crowdfunding.java
@@ -14,10 +14,10 @@ public class Crowdfunding {
     @ManyToOne
     @JoinColumn(name = "users_id")
     private Users users_id;
-    private BigDecimal current_amount = BigDecimal.valueOf(0);
-    private BigDecimal goal_amount;
+    private Integer current_amount = 0;
+    private Integer goal_amount;
 
-    public Crowdfunding(Long id, String description, OffsetDateTime created_at, Users users_id, BigDecimal current_amount, BigDecimal goal_amount) {
+    public Crowdfunding(Long id, String description, OffsetDateTime created_at, Users users_id, Integer current_amount, Integer goal_amount) {
         this.id = id;
         this.description = description;
         this.created_at = created_at;
@@ -29,7 +29,7 @@ public class Crowdfunding {
     public Crowdfunding() {
 
     }
-    public Crowdfunding(String description, Users users_id, BigDecimal goal_amount) {
+    public Crowdfunding(String description, Users users_id, Integer goal_amount) {
         this.description = description;
         this.created_at = OffsetDateTime.now();
         this.users_id = users_id;
@@ -68,19 +68,19 @@ public class Crowdfunding {
         this.users_id = users_id;
     }
 
-    public BigDecimal getCurrent_amount() {
+    public Integer getCurrent_amount() {
         return current_amount;
     }
 
-    public void setCurrent_amount(BigDecimal current_amount) {
+    public void setCurrent_amount(Integer current_amount) {
         this.current_amount = current_amount;
     }
 
-    public BigDecimal getGoal_amount() {
+    public Integer getGoal_amount() {
         return goal_amount;
     }
 
-    public void setGoal_amount(BigDecimal goal_amount) {
+    public void setGoal_amount(Integer goal_amount) {
         this.goal_amount = goal_amount;
     }
 }
